fix: Separates tests for logging in.
