+++ test/steps/application.py
@@ -1,4 +1,3 @@
-from main import Login
 from src.app import App
 
 from PyQt5.QtTest import QTest
@@ -8,43 +7,6 @@ from behave import use_step_matcher, given, when, then, step
 
 from test.steps.helpers import assert_equal_wrapper as assert_
 use_step_matcher("re")
-
-
-@when(r'I open the login dialog')
-def open_login_window(context):
-    context.login_window = Login()
-    context.login_window.show()
-
-
-@when(r'I enter "(.*)" in the (username|password) input box')
-def enter_login_input_text(context, text, dialog_box):
-    if dialog_box == "username":
-        context.login_window.login.username.setText(text)
-    elif dialog_box == "password":
-        context.login_window.login.pwd.setText(text)
-
-
-@then(r'the (username|password) input box text should be "(.*)"')
-def verify_login_input_text(context, dialog_box, text):
-    if dialog_box == "username":
-        assert context.login_window.login.username.text() == text
-    elif dialog_box == "password":
-        assert context.login_window.login.pwd.text() == text
-
-
-@when(r'I click the login button')
-def click_the_button(context, ):
-    context.login_window.login.login_button.click()
-
-
-@then(r'the login hint text should be "(.*)"')
-def verify_login_input_text(context, text):
-    assert_(context.login_window.login.login_hint.text(), text)
-
-
-@then(r'I should be able to log in to the application')
-def verify_can_log_in(context):
-    assert context.login_window.exec_()
 
 
 @when(r'I open the application window')
